Create new file named Model.php.
In this new file, create a Model class with

An attributes property (array) that is not visible outside of the class
A magic setter to create key/value pairs in the attributes array.
A magic getter to retrieve values from the attributes array based on the key name, provided the key exists.
Test your new Model class by adding key/value pairs and retrieving them.